requirements.txt
wget
tgcrypto
ffmpeg-python
py-tgcalls
telethon==1.28.2
pyrogram==2.0.59
youtube-search
youtube-search-python
yt-dlp
python-dotenv
dnspython
gitpython
asgiref
aiofiles
aiohttp
requests
pillow
heroku3
motor
psutil
future
wget
speedtest-cli
pymongo[srv]
pyromod~=1.5
googletrans==4.0.0-rc1
